LGR support: handle partial completions for CCM
Extend the retry loop for verification pattern mismatches to handle partial
completions where CPACF only processed a partial number of data blocks.
As input and output buffers may overlap (in place encryption/decryption),
only retry data that is not processed so far.

This process is only needed when having input and output buffers that might
overlap. It is not necessary when e.g. calling the KMAC instruction, where
there is no possibly overlapping output buffer. It is also not necessary when
calling the KMA instruction with only one block or less. Here, we can always
retry with the whole input data as already implemented.

The only place in the CCM implementation where we have possibly overlapping
buffers, is in routine __aes_ccm_ctr(). Code is added here to locally handle
partial completions. The overall retry loop in the encrypt and decrypt
routines is unchanged and handles all other WKVP mismatches as already
implemented.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>