LGR support: handle partial completions for GCM
Extend the retry loop for verification pattern mismatches to handle partial
completions where CPACF only processed a partial number of data blocks.
As input and output buffers may overlap (in place encryption/decryption),
only retry data that is not processed so far.

This process is only needed when having input and output buffers that might
overlap. The KMA instruction processes AAD and data simultaneously, but
an LGR could possibly interrupt AAD and/or data processing separately.
Therefore we now process AAD and data separately with two independent calls
to KMA.

Signed-off-by: Joerg Schmidbauer <jschmidb@de.ibm.com>